requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
dj-database-url==1.3.0
Django==3.2
django-allauth==0.54.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PySocks==1.7.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.3
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
